Allow to use pre-downloaded file

+++ mmsdk/mmdatasdk/computational_sequence/download_ops.py
@@ -15,8 +15,6 @@ def readURL(url,destination):
 		os.mkdir(destination.rsplit('/',1)[-2])
 	if destination is None:
 		log.error("Destination is not specified when downloading data",error=True)
-	if(os.path.isfile(destination)):
-		log.error("%s file already exists ..."%destination,error=True)
 	r = requests.get(url, stream=True)
 	if r.status_code != 200:
 		log.error('URL: %s does not exist'%url,error=True) 
@@ -24,13 +22,18 @@ def readURL(url,destination):
 	total_size = int(r.headers.get('content-length', 0)); 
 	block_size = 1024
 	wrote = 0 
-	with open(destination, 'wb') as f:
-		log.status("Downloading from %s to %s..."%(url,destination))
-		for data in tqdm(r.iter_content(block_size), total=math.ceil(total_size//block_size) , unit='MB', unit_scale=True,leave=False):
-			wrote = wrote  + len(data)
-			f.write(data)
-	f.close()
-	if total_size != 0 and wrote != total_size:
-		log.error("Error downloading the data ...")
+	if(os.path.isfile(destination)):
+		wrote = os.path.getsize(destination)
+		if wrote != total_size:
+			log.error("%s file already exists and size is different from server ..."%destination,error=True)
+	else:
+		with open(destination, 'wb') as f:
+			log.status("Downloading from %s to %s..."%(url,destination))
+			for data in tqdm(r.iter_content(block_size), total=math.ceil(total_size//block_size) , unit='MB', unit_scale=True,leave=False):
+				wrote = wrote  + len(data)
+				f.write(data)
+		f.close()
+		if total_size != 0 and wrote != total_size:
+			log.error("Error downloading the data ...")
 	log.success("Download complete!")
 	return True
